helper/cloud_accounts_pages_helper/requirements.txt
# Accounts Pages Helper - Supabase → BigQuery
# Dependências para sincronização de accounts_pages

google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
requests>=2.28.0
pandas>=2.0.0
pytz>=2023.0
pyarrow>=14.0.0
db-dtypes>=1.0.0
